change focal to float

+++ opt.py
@@ -81,7 +81,7 @@ def get_opts():
     parser.add_argument('--bounds', nargs="+", type=float, default=[2.0, 6.0],
                         help='near/far bounds for generating rays')
     
-    parser.add_argument('--focal', type=int, default=517,
+    parser.add_argument('--focal', type=float, default=517.0,
                         help='focal length')
 
     return parser.parse_args()
